Python/use_case/TCPServer/TLSTCPServer/pyproject.toml
[tool.poetry]
name = "tlstcpserver"
version = "0.1.0"
description = ""
authors = ["233Mac <ayusummer233@vip.qq.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
pyopenssl = "^25.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
